feat(neon): split instant provisioning into a `neon-instant` connector
The `neon` connector is now SDK-only: it talks to @neondatabase/serverless
and requires a connection string, so it no longer drags `neon-new` into the
dependency graph of everyone who just wants to connect to an existing Neon
database.

Auto-provisioning moves to a new `neon-instant` connector, which resolves a
connection string via `neon-new` when none is given (outside production) and
otherwise behaves identically. Shared client/statement logic lives in
`_internal/neon.ts`, which both connectors delegate to.

`neverGenerateConnectionString` is dropped, as picking the connector is now
the choice it used to express. Provisioning options (`seed`, `dotEnvFile`,
`referrer`, ...) are now actually forwarded to `instantPostgres` — they were
previously accepted in the type but ignored.
