+++ docs/2.connectors/1.index.md
@@ -11,6 +11,8 @@ Currently supported connectors:
 - [Bun](/connectors/bun)
 - [Cloudflare D1](/connectors/cloudflare)
 - [LibSQL](/connectors/libsql)
+- [Neon](/connectors/neon)
+- [Neon Instant](/connectors/neon-instant)
 - [PlanetScale](/connectors/planetscale)
 - [PostgreSQL](/connectors/postgresql)
 - [Prisma Postgres](/connectors/postgresql#prisma-postgres)
