+++ src/connectors/neon.ts
@@ -1,112 +1,12 @@
-import * as pg from "@neondatabase/serverless";
-import type { Connector, Primitive } from "db0";
+import type * as pg from "@neondatabase/serverless";
+import type { Connector } from "db0";
 
-import { BoundableStatement } from "./_internal/statement.ts";
-import { instantPostgres, type InstantPostgresParams } from "neon-new";
+import { createNeonConnector } from "./_internal/neon.ts";
 
-export type ConnectorOptions = ({ url?: string } | pg.ClientConfig) &
-  InstantPostgresParams & { neverGenerateConnectionString?: boolean };
-
-type InternalQuery = (
-  sql: string,
-  params?: Primitive[],
-) => Promise<pg.QueryResult>;
+export type ConnectorOptions = { url?: string } | pg.ClientConfig;
 
 export default function neonConnector(
   opts?: ConnectorOptions,
 ): Connector<pg.Client> {
-  let _client: undefined | pg.Client | Promise<pg.Client>;
-
-  async function getConnectionString() {
-    if (opts && "url" in opts) {
-      return opts.url;
-    } else if (opts && "connectionString" in opts) {
-      return opts.connectionString;
-    } else if (
-      process.env.NODE_ENV !== "production" &&
-      !(opts || {}).neverGenerateConnectionString
-    ) {
-      const { databaseUrl } = await instantPostgres({
-        referrer: "db0/neon-driver",
-      });
-
-      return databaseUrl;
-    }
-
-    throw new Error(
-      "[db0]:: Missing connection string for connector. Check your environment variables.",
-    );
-  }
-
-  async function getClient() {
-    if (_client) {
-      return _client;
-    }
-
-    const connectionString = await getConnectionString();
-
-    const client =
-      typeof opts === "object"
-        ? new pg.Client({ ...opts, connectionString })
-        : new pg.Client(connectionString);
-
-    _client = client.connect().then(() => {
-      _client = client;
-      return _client;
-    });
-
-    return _client;
-  }
-
-  const query: InternalQuery = async (sql, params) => {
-    const client = await getClient();
-    return client.query(normalizeParams(sql), params);
-  };
-
-  return {
-    name: "neon",
-    dialect: "postgresql",
-    getInstance: () => getClient(),
-    exec: (sql) => query(sql),
-    prepare: (sql) => new StatementWrapper(sql, query),
-    dispose: async () => {
-      await (await _client)?.end?.();
-      _client = undefined;
-    },
-  };
-}
-
-// https://www.postgresql.org/docs/9.3/sql-prepare.html
-function normalizeParams(sql: string) {
-  let i = 0;
-  return sql.replace(/\?/g, () => `$${++i}`);
-}
-
-class StatementWrapper extends BoundableStatement<void> {
-  #query: InternalQuery;
-  #sql: string;
-
-  constructor(sql: string, query: InternalQuery) {
-    super();
-    this.#sql = sql;
-    this.#query = query;
-  }
-
-  async all(...params: Primitive[]) {
-    const res = await this.#query(this.#sql, params);
-    return res.rows;
-  }
-
-  async run(...params: Primitive[]) {
-    const res = await this.#query(this.#sql, params);
-    return {
-      success: true,
-      ...res,
-    };
-  }
-
-  async get(...params: Primitive[]) {
-    const res = await this.#query(this.#sql, params);
-    return res.rows[0];
-  }
+  return createNeonConnector("neon", opts);
 }
